refactor: util directory change (#65)
