+++ src/components/common/Divider/Divider.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/className";
 
 const dividerVariants = cva("", {
   variants: {
