+++ src/components/common/Input/Input.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/className";
 
 export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {}
