+++ src/components/common/Sheet/Sheet.tsx
@@ -7,7 +7,7 @@ import { cva, type VariantProps } from "class-variance-authority";
 import * as SheetPrimitive from "@radix-ui/react-dialog";
 import * as VisuallyHidden from "@radix-ui/react-visually-hidden";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/className";
 
 const Sheet = SheetPrimitive.Root;
