+++ src/components/common/TagButton/TagButton.tsx
@@ -4,7 +4,7 @@ import React from "react";
 
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/className";
 
 const tagButtonVariants = cva(
   "rounded-lg inline-flex justify-center items-center gap-[8px] whitespace-nowrap h-[40px] px-[12px] py-[10px] border border-[#d9d9d9]",
